@@ -1,44 +1,36 @@
 ext {
     android = [
-            compileSdkVersion: 27,
-            buildToolsVersion: "27.0.2",
-            applicationId    : "com.goldze.mvvmhabit",
+            compileSdkVersion: 28,
+            buildToolsVersion: "28.0.2",
             minSdkVersion    : 16,
-            targetSdkVersion : 27,
+            targetSdkVersion : 28,
             versionCode      : 1,
             versionName      : "1.0",
     ]
     //version配置
     versions = [
-            "support-version"  : "27.1.1",
             "lifecycle-version": "1.1.1",
-            "room-version": "1.0.0",
+            "room-version"     : "1.0.0",
     ]
     //support配置
     support = [
-            "constraint-layout"       : "com.android.support.constraint:constraint-layout:1.1.0",
-            'support-v4'              : "com.android.support:support-v4:${versions["support-version"]}",
-            'appcompat-v7'            : "com.android.support:appcompat-v7:${versions["support-version"]}",
-            'recyclerview-v7'         : "com.android.support:recyclerview-v7:${versions["support-version"]}",
-            'support-v13'             : "com.android.support:support-v13:${versions["support-version"]}",
-            'support-fragment'        : "com.android.support:support-fragment:${versions["support-version"]}",
-            'design'                  : "com.android.support:design:${versions["support-version"]}",
-            'animated-vector-drawable': "com.android.support:animated-vector-drawable:${versions["support-version"]}",
-            'annotations'             : "com.android.support:support-annotations:${versions["support-version"]}",
-            'exifinterface'           : "com.android.support:exifinterface:${versions["support-version"]}",
-            'junit'                   : "junit:junit:${versions["junit-version"]}",
+            "constraint-layout": 'androidx.constraintlayout:constraintlayout:1.1.3',
+            'appcompat'        : 'androidx.appcompat:appcompat:1.0.0',
+            'recyclerview'     : 'androidx.recyclerview:recyclerview:1.0.0',
+            'design'           : 'com.google.android.material:material:1.0.0',
+            'annotations'      : 'androidx.annotation:annotation:1.0.0',
+            'exifinterface'    : 'androidx.exifinterface:exifinterface:1.0.0',
+            'fragment'         : 'androidx.fragment:fragment:1.0.0',
+            'junit'            : "junit:junit:${versions["junit-version"]}",
     ]
     //依赖第三方配置
     dependencies = [
-            //rxjava
+            //rx
             "rxjava"                               : "io.reactivex.rxjava2:rxjava:2.1.12",
             "rxandroid"                            : "io.reactivex.rxjava2:rxandroid:2.0.2",
-            //rx管理View的生命周期
             "rxlifecycle"                          : "com.trello.rxlifecycle2:rxlifecycle:2.2.1",
             "rxlifecycle-components"               : "com.trello.rxlifecycle2:rxlifecycle-components:2.2.1",
-            //rxbinding
             "rxbinding"                            : "com.jakewharton.rxbinding2:rxbinding:2.1.1",
-            //rx 6.0权限请求
             "rxpermissions"                        : "com.tbruyelle.rxpermissions2:rxpermissions:0.9.5@aar",
             //network
             "okhttp"                               : "com.squareup.okhttp3:okhttp:3.8.0",
