@@ -1,6 +1,6 @@
 package cn.finalteam.rxgalleryfinal.rxjob;
 
-import android.support.annotation.NonNull;
+import androidx.annotation.NonNull;
 
 /**
  * Desction:
