@@ -23,7 +23,7 @@ public final class FileProviderUtils {
     }
 
     private static Uri getUriForFile24(Context context, File file) {
-        Uri fileUri = android.support.v4.content.FileProvider.getUriForFile(context,
+        Uri fileUri = androidx.core.content.FileProvider.getUriForFile(context,
                 context.getPackageName() + ".fileProvider", file);
         return fileUri;
     }
