@@ -2,8 +2,9 @@ package com.asa.meta.helpers.test;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.annotation.Nullable;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.annotation.Nullable;
+import androidx.appcompat.app.AppCompatActivity;
 
 public abstract class TestActivity extends AppCompatActivity {
     public final String TAG = getClass().getSimpleName();
