@@ -2,9 +2,9 @@ package com.asa.meta.metaparty.view.activity;
 
 import android.os.Bundle;
 
-import com.android.databinding.library.baseAdapters.BR;
 import com.asa.meta.basehabit.base.LanguageActivity;
 import com.asa.meta.helpers.language.LocalManageUtil;
+import com.asa.meta.metaparty.BR;
 import com.asa.meta.metaparty.R;
 import com.asa.meta.metaparty.databinding.ActivitySwitchLanguageBinding;
 import com.asa.meta.metaparty.viewmodel.SwitchLanguageModel;
@@ -13,6 +13,7 @@ public class SwitchLanguageActivity extends LanguageActivity<ActivitySwitchLangu
 
     @Override
     public int initContentView(Bundle savedInstanceState) {
+
         return R.layout.activity_switch_language;
     }
 
