@@ -1,11 +1,12 @@
 package cn.finalteam.rxgalleryfinal.ui.adapter;
 
 import android.os.Build;
-import android.support.v4.view.PagerAdapter;
 import android.util.SparseArray;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.AdapterView;
+
+import androidx.viewpager.widget.PagerAdapter;
 
 /**
  * Desction:
