@@ -1,8 +1,9 @@
 package cn.finalteam.rxgalleryfinal.ui.widget;
 
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
 import android.view.ViewGroup;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 
 /**
