@@ -2,8 +2,9 @@ package cn.finalteam.rxgalleryfinal.ui.widget;
 
 import android.content.Context;
 import android.graphics.Rect;
-import android.support.v7.widget.RecyclerView;
 import android.view.View;
+
+import androidx.recyclerview.widget.RecyclerView;
 
 import cn.finalteam.rxgalleryfinal.R;
 
